code/backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
openai==1.54.3
psycopg2-binary==2.9.11
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
